requirements.txt
flake8
intake-dcat
intake-geopandas
jupyterlab>=1.0a3
google-cloud-bigquery
sqlalchemy<1.4.0  
ibis>=1.4.0
ibis-framework>=1.4.0 
ibis-framework[bigquery] 
ibis-bigquery>=0.1.1
python-dotenv
sendgrid